Rename timeline metrics -> metrics as it's more generic.
